+++ src/refactor_nrepl/config.clj
@@ -34,13 +34,24 @@
    :libspec-whitelist ["^cljsjs"]
 
    ;; Regexes matching paths that are to be ignored
-   :ignore-paths []})
+   :ignore-paths []
+
+   ;; Will be forwarded to clojure.pprint/*print-right-margin* when pprinting ns forms.
+   ;; You can set it to nil for disabling line wrapping.
+   ;; See also: :print-miser-width
+   :print-right-margin 72
+
+   ;; Will be forwarded to clojure.pprint/*print-miser-width* when pprinting ns forms.
+   ;; You can set it to nil for disabling line wrapping.
+   ;; See also: :print-right-margin
+   :print-miser-width 40})
 
 (defn opts-from-msg [msg]
   (into {}
         (map (fn [[k v]] (cond
                            (and (string? v) (= v "true")) [k true]
                            (and (string? v) (= v "false")) [k false]
+                           (and (string? v) (= v "nil")) [k nil]
                            :else [k v]))
              (update (select-keys msg (keys *config*))
                      :ignore-paths